refactor: extract open api logic
